Remove 'dumping?' check from footer
Moved the check to the actual call of this function, so it is never called
if we are dumping, instead of the pervious case where it was called and then
immediately returned.

Will later do the same when we are importing, which will probably have it's
very own footer anyway.

Signed-off-by: Grant Ramsay <seapagan@gmail.com>