Stop travis testing branches of the website.
By default, Travis CI will not test the gh-pages website branch. However,
it WILL test branches of this which will then fail since they are not Ruby
projects and have no Rakefile. The .yml file will ensure that only the
'master' branch will be tested.

Note that this will not be migrated to the master code branch, since we
definately DO want to test all code branches.

Signed-off-by: Grant Ramsay <seapagan@gmail.com>