Refactor the whole repo update logic.
Now the script will not change directory before each git operation, we instead
take advantage of the Git '-C' flag that allows us to perform the operation on
a repo in a completely different directory. Note however that this needs Git
version >= 1.8.5. We still need to check for this at script start up and fail
gracefully if the version is too low. Should not be too much of an issue since
most systems have a more recent version.
As a result of this, we no longer get thread race conditions where the local
directory may have been changed by a separate thread causing erroneus directories
to be reported in the fail list.

This commit also fixes a bug introduced by a previous commit that caused the
--dump and --dump-remotes to fail.

Signed-off-by: Grant Ramsay <seapagan@gmail.com>