Remove threading and consolidate Git output handling.
The use of peopen3 and threads was leading to undependable results - different
runs would have differing results dueo to race issues and similar.
Switched to popen2e which combines stdout and stderr into one stream seems to
fix this. As a bonus, some of the Git output that had previously been missed is
now also displayed.
Also consolidated the output handling into one function 'handle_output' instead
of splitting the streams.
This finally makes the issue reported in #24 100% reproducable!

Signed-off-by: Grant Ramsay <seapagan@gmail.com>