Changed folder structure, move into separate web and docs folders.
The website will be merged back into the main repository to simplify maintenance.

Website source will be placed in the web/ subdirectory while generated output will be put in the docs/ subdir.
GitHub pages can be set to serve from the docs/ subdirectory of the main branch instead of the current setting
of being on a separate gh-pages branch.

Signed-off-by: Grant Ramsay <seapagan@gmail.com>