Web: Precompile Mustache Templates.
Instead of relying upon JavaScript to compile the Mustache templates at run time,
we now pre compile these as part of the Gulp task. This results in a slightly larger
index.html, but no overhead at run time, nor missing info if JS was disabled by the user.

Also added a 'build' task that will do the same as the default task but not run the watch or
browsersync tasks.

Signed-off-by: Grant Ramsay <seapagan@gmail.com>