Fix errors in the metrics summary.
There were cases when the sum of metrics displayed in the summary did not
add up to the correct number of total repositorys. This was because the
regex used to determine if a repo was updated only matched 7 char SHA's
while some repos have 11 or more digits so were not being matched. Updated
the regex to work for 6+ digits.
Also fixed an edge case when both the updated and unchanged cases would be
matched which likewise left an error in the count.
In future we need to revisit the methods used to detect repo status

Signed-off-by: Grant Ramsay <seapagan@gmail.com>