Handle 'warning' status in Git
If a repo update generates a warning (non fatal) then it will be flagged
in a light magenta color. The summary at end of run will now display a
count of warnings along with the usual metrics.

As part of this process, the handle_output function was refactored, and
an error affecting the final count of updated repos was fixed.

Signed-off-by: Grant Ramsay <seapagan@gmail.com>